refactor: 상수 파일 분리
- 범위에 관련된 상수를 range.js 파일로 이동
