+++ src/constant/range.js
@@ -1,0 +1,13 @@
+const MIN = Object.freeze({
+  CAR_COUNT: 2,
+  RACE_COUNT: 1,
+  NAME_LENGTH: 1,
+  MOVE_CONDITION: 4,
+});
+
+const MAX = Object.freeze({
+  NAME_LENGTH: 5,
+  RANDOM_NUMBER: 9,
+});
+
+export { MIN, MAX };
