+++ src/util/random.js
@@ -1,4 +1,4 @@
-import { MAX } from "../constant/constant.js";
+import { MAX } from "../constant/range.js";
 
 const randomNumber = () => {
   return Math.floor(Math.random() * MAX.RANDOM_NUMBER);
